requirements.txt
asgiref==3.8.1
Django==5.1
sqlparse==0.5.1
djangorestframework==3.14.0
django-filter==23.2
requests==2.26.0